requirements.txt
Adafruit-Blinka==8.15.0
Adafruit-BNO055==1.0.2
adafruit-circuitpython-bno055==5.4.8
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-lis3dh==5.1.22
adafruit-circuitpython-mcp3xxx==1.4.14
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.12.12
adafruit-circuitpython-typing==1.8.3
Adafruit-GPIO==1.0.4
Adafruit-PlatformDetect==3.40.2
Adafruit-PureIO==1.1.10
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==21.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
catkin-pkg-modules==0.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.4.2
colcon-cd==0.1.1
colcon-cmake==0.2.27
colcon-common-extensions==0.3.0
colcon-core==0.12.0
colcon-defaults==0.2.8
colcon-devtools==0.2.3
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.12
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.2.4
colcon-pkg-config==0.1.0
colcon-powershell==0.3.7
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.1
colcon-ros==0.3.23
colcon-test-result==0.3.8
colcon-zsh==0.4.0
colorama==0.4.4
colorzero==2.0
command-not-found==0.3
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
duplicity==0.8.21
empy==3.3.4
entrypoints==0.4
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gattlib==0.20201113
gpio==1.0.0
gpiozero==1.6.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lgpio==0.1.0.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
nose2==0.9.2
notebook==6.5.2
notebook_shim==0.2.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
PAM==0.4.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
PyBluez-updated==0.31
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.17
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
pyftdi==0.54.0
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.19.2
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pythran==0.10.0
pytz==2022.1
pyusb==1.2.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.8
requests==2.25.1
roman==3.3
rosdistro-modules==0.9.0
rospkg-modules==1.4.0
RPi.GPIO==0.7.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.1
spidev==3.5
ssh-import-id==5.11
stack-data==0.6.0
sympy==1.9
systemd-python==234
terminado==0.17.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0
